Switch features to uint_8t (#547)
For initial token use, we're going to re-use feature offsets to be
feature ids, so sizing these correctly for that purpose.

---------

+++ mettagrid/benchmarks/grid_object_benchmark.cpp
@@ -14,7 +14,7 @@ public:
     init(type_id, r, c, layer);
   }
 
-  void obs(ObsType* obs, const std::vector<unsigned int>& offsets) const override {
+  void obs(ObsType* obs, const std::vector<uint8_t>& offsets) const override {
     // Simple implementation for benchmarking
     for (size_t i = 0; i < offsets.size(); ++i) {
       obs[offsets[i]] = i % 255;
@@ -75,7 +75,7 @@ static void BM_GridObjectObs(benchmark::State& state) {
   TestGridObject obj(1, 5, 10, 0);
 
   // Create offsets vector and observation buffer
-  std::vector<unsigned int> offsets(numOffsets);
+  std::vector<uint8_t> offsets(numOffsets);
   for (int i = 0; i < numOffsets; ++i) {
     offsets[i] = i;
   }
